drivers/kw41zrf: fix incorrect rssi for last packet received
Apparently this value is given by the hardware as an int8_t in dBm directly
although I can't find where the datasheet says this explicitly.

What's also not clear is whether or not this rssi is specifically for the
last packet received. The lqi in the same register is.

https://community.nxp.com/thread/385126#comment-613185

+++ drivers/kw41zrf/kw41zrf_netdev.c
@@ -330,7 +330,7 @@ static int kw41zrf_netdev_recv(netdev_t *netdev, void *buf, size_t len, void *in
         } else {
             radio_info->lqi = (51 * hw_lqi) / 44;
         }
-        radio_info->rssi = (ZLL->LQI_AND_RSSI & ZLL_LQI_AND_RSSI_RSSI_MASK) >> ZLL_LQI_AND_RSSI_RSSI_SHIFT;
+        radio_info->rssi = (int8_t)((ZLL->LQI_AND_RSSI & ZLL_LQI_AND_RSSI_RSSI_MASK) >> ZLL_LQI_AND_RSSI_RSSI_SHIFT);
     }
 
     /* Go back to RX mode */
